+++ kwcoco/coco_dataset.py
@@ -4030,6 +4030,20 @@ class CocoDataset(ub.NiceRepr, MixinCocoAddRemove, MixinCocoStats,
             else:
                 json.dump(self.dataset, file, indent=indent, ensure_ascii=False)
 
+    def _check_json_serializable(self, verbose=1):
+        """
+        Debug which part of a coco dataset might not be json serializable
+        """
+        from kwcoco.util.util_json import find_json_unserializable
+        bad_parts = find_json_unserializable(self.dataset)
+
+        if verbose:
+            if bad_parts:
+                print(ub.repr2(bad_parts))
+            summary = 'There are {} total errors'.format(len(bad_parts))
+            print('summary = {}'.format(summary))
+        return bad_parts
+
     def _check_integrity(self):
         """ perform all checks """
         self._check_index()